Split component, template, styles, and functionality
